:fire: useless functionality and improve structure
