+++ main.go
@@ -1,10 +1,12 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
 
-	"github.com/ivanilves/lstags/app"
+	"github.com/jessevdk/go-flags"
+
 	"github.com/ivanilves/lstags/auth"
 	"github.com/ivanilves/lstags/docker"
 	dockerclient "github.com/ivanilves/lstags/docker/client"
@@ -12,16 +14,62 @@ import (
 	"github.com/ivanilves/lstags/tag"
 	"github.com/ivanilves/lstags/tag/local"
 	"github.com/ivanilves/lstags/tag/registry"
+	"github.com/ivanilves/lstags/util"
 )
 
-var doNotFail bool
+// Options represents configuration options we extract from passed command line arguments
+type Options struct {
+	DockerJSON         string `short:"j" long:"docker-json" default:"~/.docker/config.json" description:"JSON file with credentials" env:"DOCKER_JSON"`
+	Pull               bool   `short:"P" long:"pull" description:"Pull Docker images matched by filter (will use local Docker deamon)" env:"PULL"`
+	PushRegistry       string `short:"U" long:"push-registry" description:"[Re]Push pulled images to a specified remote registry" env:"PUSH_REGISTRY"`
+	PushPrefix         string `short:"R" long:"push-prefix" description:"[Re]Push pulled images with a specified repo path prefix" env:"PUSH_PREFIX"`
+	ConcurrentRequests int    `short:"c" long:"concurrent-requests" default:"32" description:"Limit of concurrent requests to the registry" env:"CONCURRENT_REQUESTS"`
+	TraceRequests      bool   `short:"T" long:"trace-requests" description:"Trace Docker registry HTTP requests" env:"TRACE_REQUESTS"`
+	DoNotFail          bool   `short:"N" long:"do-not-fail" description:"Do not fail on non-critical errors (could be dangerous!)" env:"DO_NOT_FAIL"`
+	Version            bool   `short:"V" long:"version" description:"Show version and exit"`
+	Positional         struct {
+		Repositories []string `positional-arg-name:"REPO1 REPO2 REPOn" description:"Docker repositories to operate on, e.g.: alpine nginx~/1\\.13\\.5$/ busybox~/1.27.2/"`
+	} `positional-args:"yes" required:"yes"`
+}
 
-func suicide(err error) {
+var doNotFail = false
+
+func suicide(err error, critical bool) {
 	fmt.Printf("%s\n", err.Error())
 
-	if !doNotFail {
+	if doNotFail || critical {
 		os.Exit(1)
 	}
+}
+
+func parseFlags() (*Options, error) {
+	var err error
+
+	o := &Options{}
+
+	_, err = flags.Parse(o)
+	if err != nil {
+		os.Exit(1) // YES! Just exit!
+	}
+
+	if o.Version {
+		fmt.Printf("VERSION: %s\n", getVersion())
+		os.Exit(0)
+	}
+
+	if len(o.Positional.Repositories) == 0 {
+		return nil, errors.New("Need at least one repository name, e.g. 'nginx~/^1\\\\.13/' or 'mesosphere/chronos'")
+	}
+
+	if o.PushRegistry != "" {
+		o.Pull = true
+	}
+
+	registry.TraceRequests = o.TraceRequests
+
+	doNotFail = o.DoNotFail
+
+	return o, nil
 }
 
 func getVersion() string {
@@ -33,30 +81,15 @@ func getAuthorization(t auth.TokenResponse) string {
 }
 
 func main() {
-	o, err := app.ParseFlags()
+	o, err := parseFlags()
 	if err != nil {
-		suicide(err)
+		suicide(err, true)
 	}
-
-	if o.Version {
-		fmt.Printf("VERSION: %s\n", getVersion())
-		os.Exit(0)
-	}
-
-	dockerconfig.DefaultUsername = o.Username
-	dockerconfig.DefaultPassword = o.Password
 
 	dockerConfig, err := dockerconfig.Load(o.DockerJSON)
 	if err != nil {
-		suicide(err)
+		suicide(err, true)
 	}
-
-	registry.TraceRequests = o.TraceRequests
-
-	auth.WebSchema = o.GetWebSchema()
-	registry.WebSchema = o.GetWebSchema()
-
-	doNotFail = o.DoNotFail
 
 	const format = "%-12s %-45s %-15s %-25s %s\n"
 	fmt.Printf(format, "<STATE>", "<DIGEST>", "<(local) ID>", "<Created At>", "<TAG>")
@@ -65,11 +98,9 @@ func main() {
 	pullCount := 0
 	pushCount := 0
 
-	pullAuths := make(map[string]string)
-
 	dc, err := dockerclient.New(dockerConfig)
 	if err != nil {
-		suicide(err)
+		suicide(err, true)
 	}
 
 	type tagResult struct {
@@ -82,10 +113,10 @@ func main() {
 	trc := make(chan tagResult, repoCount)
 
 	for _, r := range o.Positional.Repositories {
-		go func(r string, o *app.Options, trc chan tagResult) {
-			repository, filter, err := app.SeparateFilterAndRepo(r)
+		go func(r string, o *Options, trc chan tagResult) {
+			repository, filter, err := util.SeparateFilterAndRepo(r)
 			if err != nil {
-				suicide(err)
+				suicide(err, true)
 			}
 
 			registryName := docker.GetRegistry(repository)
@@ -95,27 +126,25 @@ func main() {
 
 			username, password, _ := dockerConfig.GetCredentials(registryName)
 
-			pullAuths[repoLocalName], _ = dockerConfig.GetRegistryAuth(registryName)
-
 			tresp, err := auth.NewToken(registryName, repoRegistryName, username, password)
 			if err != nil {
-				suicide(err)
+				suicide(err, true)
 			}
 
 			authorization := getAuthorization(tresp)
 
 			registryTags, err := registry.FetchTags(registryName, repoRegistryName, authorization, o.ConcurrentRequests)
 			if err != nil {
-				suicide(err)
+				suicide(err, true)
 			}
 
 			imageSummaries, err := dc.ListImagesForRepo(repoLocalName)
 			if err != nil {
-				suicide(err)
+				suicide(err, true)
 			}
 			localTags, err := local.FetchTags(repoLocalName, imageSummaries)
 			if err != nil {
-				suicide(err)
+				suicide(err, true)
 			}
 
 			sortedKeys, names, joinedTags := tag.Join(registryTags, localTags)
@@ -126,7 +155,7 @@ func main() {
 
 				tg := joinedTags[name]
 
-				if !app.DoesMatch(tg.GetName(), filter) {
+				if !util.DoesMatch(tg.GetName(), filter) {
 					continue
 				}
 
@@ -177,7 +206,7 @@ func main() {
 						fmt.Printf("PULLING %s\n", ref)
 						err := dc.Pull(ref)
 						if err != nil {
-							suicide(err)
+							suicide(err, false)
 						}
 
 						done <- true
@@ -207,7 +236,7 @@ func main() {
 				for _, tg := range tags {
 					prefix := o.PushPrefix
 					if prefix == "" {
-						prefix = app.GeneratePathFromHostname(registry)
+						prefix = util.GeneratePathFromHostname(registry)
 					}
 
 					srcRef := repo + ":" + tg.GetName()
@@ -217,12 +246,12 @@ func main() {
 
 					err := dc.Tag(srcRef, dstRef)
 					if err != nil {
-						suicide(err)
+						suicide(err, true)
 					}
 
 					err = dc.Push(dstRef)
 					if err != nil {
-						suicide(err)
+						suicide(err, false)
 					}
 
 					done <- true
